Updating client version information
Many of the version strings have not been updated in the
wallet source. Because any changes at this point are likely to
require a hard fork, I have bumped RPC version and RPC minimum
version to connect to 102000. This will allow the new clients
to sync to the proper blockchain (and prevent older clients from
destroying syncing).